refactor: implement lazy module loading to improve startup performance
- Added lazy import system for core modules and services to defer initialization until first access
- Introduced `_LazyCoreProxy` pattern to enable on-demand loading without blocking module import
- Updated `.pylintrc` to ignore auto-generated `.pyi` stub files
- Enhanced README documentation for startup orchestration, service initialization patterns, and infrastructure modules

+++ tests/test_backtest_optimizer_basic.py
@@ -9,7 +9,7 @@ class MockStrategy:
         self.params = params
         self.equity = []
         self.returns = []
-        
+
     def run_backtest(self, data: List[Dict]) -> None:
         """模拟回测执行"""
         # 简单的模拟策略：根据参数计算收益
@@ -27,22 +27,22 @@ def test_backtest_optimizer():
         {'open': 10.2, 'high': 10.8, 'low': 10.1, 'close': 10.5, 'volume': 1200},
         {'open': 10.5, 'high': 11.0, 'low': 10.4, 'close': 10.8, 'volume': 1500}
     ]
-    
+
     # 定义参数网格
     param_grid = {
         'leverage': [1.0, 2.0, 3.0],
         'ma_period': [5, 10, 20]
     }
-    
+
     # 创建优化器
     optimizer = BacktestOptimizer()
-    
+
     # 执行优化
     def mock_backtest(strategy_cls, data, symbol, exchange, **params):
         strategy = strategy_cls(params)
         strategy.run_backtest(data)
         return strategy.returns, strategy.equity
-    
+
     # 模拟优化过程
     results = []
     from itertools import product
@@ -50,10 +50,10 @@ def test_backtest_optimizer():
         param_dict = dict(zip(param_grid.keys(), params))
         strategy = MockStrategy(param_dict)
         strategy.run_backtest(test_data)
-        sharpe = np.mean(strategy.returns) / (np.std(strategy.returns) + 1e-8)
+        sharpe = float(np.mean(strategy.returns) / (np.std(strategy.returns) + 1e-8))
         mdd = 0.1  # 简化的最大回撤计算
         results.append(BacktestResult(param_dict, sharpe, mdd))
-    
+
     # 验证结果
     assert len(results) == 9  # 3x3 参数组合
     assert all(isinstance(r, BacktestResult) for r in results)
